+++ src/core/errors.js
@@ -1,15 +1,10 @@
 'use strict';
 
 /**
- * Custom error classes following clean code principles
- * Provides clear, actionable error messages with context
+ * Base class for all CSV parsing errors.
+ * Provides consistent error formatting and context.
  * @category Error Classes
- */
-
-/**
- * Base class for all CSV parsing errors
- * Provides consistent error formatting and context
- * @category Error Classes
+ * @augments Error
  */
 class CsvParsingError extends Error {
     /**
@@ -59,9 +54,10 @@ class CsvParsingError extends Error {
 }
 
 /**
- * Input validation errors
- * Thrown when function parameters don't meet expected type or value requirements
+ * Input validation errors.
+ * Thrown when function parameters don't meet expected type or value requirements.
  * @category Error Classes
+ * @augments CsvParsingError
  */
 class InputValidationError extends CsvParsingError {
     /**
@@ -87,9 +83,10 @@ class InputValidationError extends CsvParsingError {
 }
 
 /**
- * Configuration-related errors
- * Thrown when configuration options conflict or are invalid
+ * Configuration-related errors.
+ * Thrown when configuration options conflict or are invalid.
  * @category Error Classes
+ * @augments CsvParsingError
  */
 class ConfigurationError extends CsvParsingError {
     /**
@@ -143,9 +140,10 @@ class ConfigurationError extends CsvParsingError {
 }
 
 /**
- * CSV parsing errors with detailed context
- * Thrown when CSV format is invalid or malformed
+ * CSV parsing errors with detailed context.
+ * Thrown when CSV format is invalid or malformed.
  * @category Error Classes
+ * @augments CsvParsingError
  */
 class CsvFormatError extends CsvParsingError {
     /**
@@ -203,9 +201,10 @@ class CsvFormatError extends CsvParsingError {
 }
 
 /**
- * File operation errors
- * Thrown when file read or write operations fail
+ * File operation errors.
+ * Thrown when file read or write operations fail.
  * @category Error Classes
+ * @augments CsvParsingError
  */
 class FileOperationError extends CsvParsingError {
     /**
@@ -236,9 +235,10 @@ class FileOperationError extends CsvParsingError {
 }
 
 /**
- * JSON validation errors
- * Thrown when parsed CSV data cannot be converted to valid JSON
+ * JSON validation errors.
+ * Thrown when parsed CSV data cannot be converted to valid JSON.
  * @category Error Classes
+ * @augments CsvParsingError
  */
 class JsonValidationError extends CsvParsingError {
     /**
@@ -267,9 +267,10 @@ class JsonValidationError extends CsvParsingError {
 }
 
 /**
- * Browser-specific errors
- * Thrown when browser API operations fail
+ * Browser-specific errors.
+ * Thrown when browser API operations fail.
  * @category Error Classes
+ * @augments CsvParsingError
  */
 class BrowserApiError extends CsvParsingError {
     /**
