+++ src/core/parserConfig.js
